refactor(user): improve readability
