--- services/user-service/src/routes/me/avatar.js
@@ -1,3 +1,0 @@
-export default async function avatarRoute(app) {
-
-}
