--- services/user-service/src/routes/me/index.js
@@ -1,8 +1,0 @@
-import meRoute from '../me/me.js'
-import avatarRoute from './../me/avatar.js';
-
-export default async function meRoutes(app) {
-
-    app.register(meRoute);
-    app.register(avatarRoute, { prefix: '/avatar' });
-}
