--- services/user-service/src/routes/users-resources/avatar.routes.js
@@ -1,3 +1,0 @@
-export default async function avatarRoutes(app) {
-
-}
