--- services/user-service/src/routes/users-resources/index.js
@@ -1,4 +1,0 @@
-import avatarRoutes  from "./avatar.routes.js";
-export default async function userRoutes(app) {
-  await avatarRoutes(app);
-}
