+++ services/user-service/src/routes/users/me/index.js
@@ -1,0 +1,6 @@
+import meRoute from './me.routes.js'
+
+export default async function meRoutes(app) {
+
+    app.register(meRoute);
+}
